Split payment items from kansa/ to payments/
