+++ src/index.jsx
@@ -18,8 +18,8 @@ import App from './app/components/App'
 import Index from './app/components/Index'
 import NewMemberForm from './kansa/components/NewMemberForm'
 import NewMemberIndex from './kansa/components/NewMemberIndex'
-import PurchaseIndex from './kansa/components/PurchaseIndex'
-import PurchaseItem from './kansa/components/PurchaseItem'
+import PurchaseIndex from './payments/components/PurchaseIndex'
+import PurchaseItem from './payments/components/PurchaseItem'
 import Nominate from './hugo/components/Nominate'
 import Canon from './hugo-admin/components/Canon'
 import Finalists from './hugo-admin/components/Finalists'
