+++ src/reducers.js
@@ -4,7 +4,7 @@ import { routerReducer } from 'react-router-redux'
 import app from './app/reducer'
 import hugoAdmin from './hugo-admin/reducer'
 import nominations from './hugo/reducers/nominations'
-import purchase from './kansa/reducers/purchase'
+import purchase from './payments/reducers'
 import user from './kansa/reducers/user'
 
 export default combineReducers({
